requirements.txt
cryptography==3.4.7
Flask==2.0.1
passlib==1.7.4
pymongo==3.11.4
python-dotenv==0.18.0
